requirements.txt
streamlit
requests
python-dotenv
langchain
langchain-community
sentence-transformers
faiss-cpu